requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
pypdf
sentence-transformers